+++ src/watchdog/events.py
@@ -105,6 +105,8 @@ from __future__ import annotations
 import logging
 import os.path
 import re
+from dataclasses import dataclass, field
+from typing import Optional
 
 from watchdog.utils.patterns import match_any_paths
 
@@ -117,6 +119,7 @@ EVENT_TYPE_OPENED = "opened"
 EVENT_TYPE_ATTRIB = "attrib"
 
 
+@dataclass(unsafe_hash=True)
 class FileSystemEvent:
     """
     Immutable type that represents a file system event that is triggered
@@ -126,80 +129,23 @@ class FileSystemEvent:
     can be used as keys in dictionaries or be added to sets.
     """
 
-    event_type = ""
-    """The type of the event as a string."""
+    src_path: str
+    dest_path: str = ""
+    event_type: str = field(default="", init=False)
+    is_directory: bool = field(default=False, init=False)
 
-    is_directory = False
-    """True if event was emitted for a directory; False otherwise."""
-
-    is_synthetic = False
     """
     True if event was synthesized; False otherwise.
-
     These are events that weren't actually broadcast by the OS, but
     are presumed to have happened based on other, actual events.
     """
-
-    def __init__(self, src_path):
-        self._src_path = src_path
-
-    @property
-    def src_path(self):
-        """Source path of the file system object that triggered this event."""
-        return self._src_path
-
-    def __str__(self):
-        return self.__repr__()
-
-    def __repr__(self):
-        return (
-            f"<{type(self).__name__}: event_type={self.event_type}, "
-            f"src_path={self.src_path!r}, is_directory={self.is_directory}"
-        )
-
-    # Used for comparison of events.
-    @property
-    def key(self):
-        return (self.event_type, self.src_path, self.is_directory)
-
-    def __eq__(self, event):
-        if isinstance(event, FileSystemEvent):
-            return self.key == event.key
-        return NotImplemented
-
-    def __ne__(self, event):
-        return self.key != event.key
-
-    def __hash__(self):
-        return hash(self.key)
+    is_synthetic: bool = field(default=False)
 
 
 class FileSystemMovedEvent(FileSystemEvent):
-    """
-    File system event representing any kind of file system movement.
-    """
+    """File system event representing any kind of file system movement."""
 
     event_type = EVENT_TYPE_MOVED
-
-    def __init__(self, src_path, dest_path):
-        super().__init__(src_path)
-        self._dest_path = dest_path
-
-    @property
-    def dest_path(self):
-        """The destination path of the move event."""
-        return self._dest_path
-
-    # Used for hashing this as an immutable object.
-    @property
-    def key(self):
-        return (self.event_type, self.src_path, self.dest_path, self.is_directory)
-
-    def __repr__(self):
-        return (
-            f"<{type(self).__name__}: src_path={self.src_path!r}, "
-            f"dest_path={self.dest_path!r}, is_directory={self.is_directory}>"
-        )
 
 
 # File events.
@@ -291,7 +237,7 @@ class FileSystemEventHandler:
     Base file system event handler that you can override methods from.
     """
 
-    def dispatch(self, event):
+    def dispatch(self, event: FileSystemEvent) -> None:
         """Dispatches events to the appropriate methods.
 
         :param event:
@@ -310,7 +256,7 @@ class FileSystemEventHandler:
             EVENT_TYPE_ATTRIB: self.on_attrib,
         }[event.event_type](event)
 
-    def on_any_event(self, event):
+    def on_any_event(self, event: FileSystemEvent) -> None:
         """Catch-all event handler.
 
         :param event:
@@ -319,7 +265,7 @@ class FileSystemEventHandler:
             :class:`FileSystemEvent`
         """
 
-    def on_moved(self, event):
+    def on_moved(self, event: FileSystemEvent) -> None:
         """Called when a file or a directory is moved or renamed.
 
         :param event:
@@ -328,7 +274,7 @@ class FileSystemEventHandler:
             :class:`DirMovedEvent` or :class:`FileMovedEvent`
         """
 
-    def on_created(self, event):
+    def on_created(self, event: FileSystemEvent) -> None:
         """Called when a file or directory is created.
 
         :param event:
@@ -337,7 +283,7 @@ class FileSystemEventHandler:
             :class:`DirCreatedEvent` or :class:`FileCreatedEvent`
         """
 
-    def on_deleted(self, event):
+    def on_deleted(self, event: FileSystemEvent) -> None:
         """Called when a file or directory is deleted.
 
         :param event:
@@ -346,7 +292,7 @@ class FileSystemEventHandler:
             :class:`DirDeletedEvent` or :class:`FileDeletedEvent`
         """
 
-    def on_modified(self, event):
+    def on_modified(self, event: FileSystemEvent) -> None:
         """Called when a file or directory is modified.
 
         :param event:
@@ -355,7 +301,7 @@ class FileSystemEventHandler:
             :class:`DirModifiedEvent` or :class:`FileModifiedEvent`
         """
 
-    def on_closed(self, event):
+    def on_closed(self, event: FileSystemEvent) -> None:
         """Called when a file opened for writing is closed.
 
         :param event:
@@ -364,7 +310,7 @@ class FileSystemEventHandler:
             :class:`FileClosedEvent`
         """
 
-    def on_opened(self, event):
+    def on_opened(self, event: FileSystemEvent) -> None:
         """Called when a file is opened.
 
         :param event:
@@ -373,8 +319,9 @@ class FileSystemEventHandler:
             :class:`FileOpenedEvent`
         """
 
-    def on_attrib(self, event):
+    def on_attrib(self, event: FileSystemEvent) -> None:
         """Called when a file or directory metadata is modified.
+
         :param event:
             Event representing file/directory metadata modification.
         :type event:
@@ -434,7 +381,7 @@ class PatternMatchingEventHandler(FileSystemEventHandler):
         """
         return self._case_sensitive
 
-    def dispatch(self, event):
+    def dispatch(self, event: FileSystemEvent) -> None:
         """Dispatches events to the appropriate methods.
 
         :param event:
@@ -522,7 +469,7 @@ class RegexMatchingEventHandler(FileSystemEventHandler):
         """
         return self._case_sensitive
 
-    def dispatch(self, event):
+    def dispatch(self, event: FileSystemEvent) -> None:
         """Dispatches events to the appropriate methods.
 
         :param event:
@@ -549,32 +496,29 @@ class RegexMatchingEventHandler(FileSystemEventHandler):
 class LoggingEventHandler(FileSystemEventHandler):
     """Logs all the events captured."""
 
-    def __init__(self, logger=None):
+    def __init__(self, logger: Optional[logging.Logger] = None) -> None:
         super().__init__()
-
         self.logger = logger or logging.root
 
-    def on_moved(self, event):
+    def on_moved(self, event: FileSystemEvent) -> None:
         super().on_moved(event)
 
         what = "directory" if event.is_directory else "file"
-        self.logger.info(
-            "Moved %s: from %s to %s", what, event.src_path, event.dest_path
-        )
+        self.logger.info("Moved %s: from %s to %s", what, event.src_path, event.dest_path)
 
-    def on_created(self, event):
+    def on_created(self, event: FileSystemEvent) -> None:
         super().on_created(event)
 
         what = "directory" if event.is_directory else "file"
         self.logger.info("Created %s: %s", what, event.src_path)
 
-    def on_deleted(self, event):
+    def on_deleted(self, event: FileSystemEvent) -> None:
         super().on_deleted(event)
 
         what = "directory" if event.is_directory else "file"
         self.logger.info("Deleted %s: %s", what, event.src_path)
 
-    def on_modified(self, event):
+    def on_modified(self, event: FileSystemEvent) -> None:
         super().on_modified(event)
 
         what = "directory" if event.is_directory else "file"
@@ -585,6 +529,16 @@ class LoggingEventHandler(FileSystemEventHandler):
 
         what = "directory" if event.is_directory else "file"
         self.logger.info("Attrib %s: %s", what, event.src_path)
+
+    def on_closed(self, event: FileSystemEvent) -> None:
+        super().on_closed(event)
+
+        self.logger.info("Closed file: %s", event.src_path)
+
+    def on_opened(self, event: FileSystemEvent) -> None:
+        super().on_opened(event)
+
+        self.logger.info("Opened file: %s", event.src_path)
 
 
 def generate_sub_moved_events(src_dir_path, dest_dir_path):
@@ -603,20 +557,12 @@ def generate_sub_moved_events(src_dir_path, dest_dir_path):
     for root, directories, filenames in os.walk(dest_dir_path):
         for directory in directories:
             full_path = os.path.join(root, directory)
-            renamed_path = (
-                full_path.replace(dest_dir_path, src_dir_path) if src_dir_path else None
-            )
-            dir_moved_event = DirMovedEvent(renamed_path, full_path)
-            dir_moved_event.is_synthetic = True
-            yield dir_moved_event
+            renamed_path = full_path.replace(dest_dir_path, src_dir_path) if src_dir_path else ""
+            yield DirMovedEvent(renamed_path, full_path, is_synthetic=True)
         for filename in filenames:
             full_path = os.path.join(root, filename)
-            renamed_path = (
-                full_path.replace(dest_dir_path, src_dir_path) if src_dir_path else None
-            )
-            file_moved_event = FileMovedEvent(renamed_path, full_path)
-            file_moved_event.is_synthetic = True
-            yield file_moved_event
+            renamed_path = full_path.replace(dest_dir_path, src_dir_path) if src_dir_path else ""
+            yield FileMovedEvent(renamed_path, full_path, is_synthetic=True)
 
 
 def generate_sub_created_events(src_dir_path):
@@ -632,10 +578,6 @@ def generate_sub_created_events(src_dir_path):
     """
     for root, directories, filenames in os.walk(src_dir_path):
         for directory in directories:
-            dir_created_event = DirCreatedEvent(os.path.join(root, directory))
-            dir_created_event.is_synthetic = True
-            yield dir_created_event
+            yield DirCreatedEvent(os.path.join(root, directory), is_synthetic=True)
         for filename in filenames:
-            file_created_event = FileCreatedEvent(os.path.join(root, filename))
-            file_created_event.is_synthetic = True
-            yield file_created_event
+            yield FileCreatedEvent(os.path.join(root, filename), is_synthetic=True)