+++ tests/markers.py
@@ -4,6 +4,4 @@ from platform import python_implementation
 
 import pytest
 
-cpython_only = pytest.mark.skipif(
-    python_implementation() != "CPython", reason="CPython only."
-)
+cpython_only = pytest.mark.skipif(python_implementation() != "CPython", reason="CPython only.")