+++ tests/test_emitter.py
@@ -53,10 +53,7 @@ if platform.is_darwin():
 
 def rerun_filter(exc, *args):
     time.sleep(5)
-    if issubclass(exc[0], Empty) and platform.is_windows():
-        return True
-
-    return False
+    return bool(issubclass(exc[0], Empty) and platform.is_windows())
 
 
 @pytest.mark.flaky(max_runs=5, min_passes=1, rerun_filter=rerun_filter)
@@ -79,9 +76,7 @@ def test_create(p: P, event_queue: TestEventQueue, start_watching: StartWatching
 
 
 @pytest.mark.xfail(reason="known to be problematic")
-@pytest.mark.skipif(
-    not platform.is_linux(), reason="FileCloseEvent only supported in GNU/Linux"
-)
+@pytest.mark.skipif(not platform.is_linux(), reason="FileCloseEvent only supported in GNU/Linux")
 @pytest.mark.flaky(max_runs=5, min_passes=1, rerun_filter=rerun_filter)
 def test_close(p: P, event_queue: TestEventQueue, start_watching: StartWatching) -> None:
     f_d = open(p("a"), "a")
@@ -254,9 +249,7 @@ def test_move_to(p: P, start_watching: StartWatching, expect_event: ExpectEvent)
         expect_event(DirModifiedEvent(p("dir2")))
 
 
-@pytest.mark.skipif(
-    not platform.is_linux(), reason="InotifyFullEmitter only supported in Linux"
-)
+@pytest.mark.skipif(not platform.is_linux(), reason="InotifyFullEmitter only supported in Linux")
 def test_move_to_full(p: P, event_queue: TestEventQueue, start_watching: StartWatching) -> None:
     mkdir(p("dir1"))
     mkdir(p("dir2"))
@@ -267,7 +260,7 @@ def test_move_to_full(p: P, event_queue: TestEventQueue, start_watching: StartWa
     event = event_queue.get(timeout=5)[0]
     assert isinstance(event, FileMovedEvent)
     assert event.dest_path == p("dir2", "b")
-    assert event.src_path is None  # Should equal None since the path was not watched
+    assert event.src_path == ""  # Should be blank since the path was not watched
 
 
 @pytest.mark.flaky(max_runs=5, min_passes=1, rerun_filter=rerun_filter)
@@ -285,9 +278,7 @@ def test_move_from(p: P, start_watching: StartWatching, expect_event: ExpectEven
         expect_event(DirModifiedEvent(p("dir1")))
 
 
-@pytest.mark.skipif(
-    not platform.is_linux(), reason="InotifyFullEmitter only supported in Linux"
-)
+@pytest.mark.skipif(not platform.is_linux(), reason="InotifyFullEmitter only supported in Linux")
 def test_move_from_full(p: P, event_queue: TestEventQueue, start_watching: StartWatching) -> None:
     mkdir(p("dir1"))
     mkdir(p("dir2"))
@@ -298,7 +289,7 @@ def test_move_from_full(p: P, event_queue: TestEventQueue, start_watching: Start
     event = event_queue.get(timeout=5)[0]
     assert isinstance(event, FileMovedEvent)
     assert event.src_path == p("dir1", "a")
-    assert event.dest_path is None  # Should equal None since path not watched
+    assert event.dest_path == ""  # Should be blank since path not watched
 
 
 @pytest.mark.flaky(max_runs=5, min_passes=1, rerun_filter=rerun_filter)
@@ -329,9 +320,7 @@ def test_separate_consecutive_moves(p: P, start_watching: StartWatching, expect_
 
 
 @pytest.mark.flaky(max_runs=5, min_passes=1, rerun_filter=rerun_filter)
-@pytest.mark.skipif(
-    platform.is_bsd(), reason="BSD create another set of events for this test"
-)
+@pytest.mark.skipif(platform.is_bsd(), reason="BSD create another set of events for this test")
 def test_delete_self(p: P, start_watching: StartWatching, expect_event: ExpectEvent) -> None:
     mkdir(p("dir1"))
     emitter = start_watching(p("dir1"))
@@ -368,9 +357,7 @@ def test_fast_subdirectory_creation_deletion(p: P, event_queue: TestEventQueue, 
         count[etype] += 1
         assert event.src_path == etype_for_dir[etype]
         assert count[DirCreatedEvent] >= count[DirDeletedEvent]
-        assert (
-            count[DirCreatedEvent] + count[DirDeletedEvent] >= count[DirModifiedEvent]
-        )
+        assert count[DirCreatedEvent] + count[DirDeletedEvent] >= count[DirModifiedEvent]
     assert count == {
         DirCreatedEvent: times,
         DirModifiedEvent: times * 2,
@@ -450,9 +437,7 @@ def test_recursive_off(
             event_queue.get(timeout=5)
 
         mkdir(p("dir3"))
-        expect_event(
-            DirModifiedEvent(p())
-        )  # the contents of the parent directory changed
+        expect_event(DirModifiedEvent(p()))  # the contents of the parent directory changed
 
         mv(p("dir1", "dir2", "somefile"), p("somefile"))
         expect_event(FileMovedEvent(p("dir1", "dir2", "somefile"), p("somefile")))
@@ -463,9 +448,7 @@ def test_recursive_off(
         expect_event(DirModifiedEvent(p()))
 
 
-@pytest.mark.skipif(
-    platform.is_windows(), reason="Windows create another set of events for this test"
-)
+@pytest.mark.flaky(max_runs=5, min_passes=1, rerun_filter=rerun_filter)
 def test_renaming_top_level_directory(
     p: P,
     event_queue: TestEventQueue,
@@ -476,7 +459,8 @@ def test_renaming_top_level_directory(
 
     mkdir(p("a"))
     expect_event(DirCreatedEvent(p("a")))
-    expect_event(DirModifiedEvent(p()))
+    if not platform.is_windows():
+        expect_event(DirModifiedEvent(p()))
 
     mkdir(p("a", "b"))
     expect_event(DirCreatedEvent(p("a", "b")))
@@ -484,10 +468,10 @@ def test_renaming_top_level_directory(
 
     mv(p("a"), p("a2"))
     expect_event(DirMovedEvent(p("a"), p("a2")))
-    expect_event(DirModifiedEvent(p()))
-    expect_event(DirModifiedEvent(p()))
-
-    expect_event(DirMovedEvent(p("a", "b"), p("a2", "b")))
+    if not platform.is_windows():
+        expect_event(DirModifiedEvent(p()))
+        expect_event(DirModifiedEvent(p()))
+    expect_event(DirMovedEvent(p("a", "b"), p("a2", "b"), is_synthetic=True))
 
     if platform.is_bsd():
         expect_event(DirModifiedEvent(p()))
@@ -502,19 +486,8 @@ def test_renaming_top_level_directory(
             break
 
     assert all(
-        [
-            isinstance(
-                e,
-                (
-                    FileCreatedEvent,
-                    FileMovedEvent,
-                    FileOpenedEvent,
-                    DirModifiedEvent,
-                    FileClosedEvent,
-                ),
-            )
-            for e in events
-        ]
+        isinstance(e, (FileCreatedEvent, FileMovedEvent, FileOpenedEvent, DirModifiedEvent, FileClosedEvent))
+        for e in events
     )
 
     for event in events:
@@ -527,74 +500,7 @@ def test_renaming_top_level_directory(
             assert event.src_path == p("a2", "b")
 
 
-@pytest.mark.flaky(max_runs=5, min_passes=1, rerun_filter=rerun_filter)
-@pytest.mark.skipif(
-    not platform.is_windows(),
-    reason="Non-Windows create another set of events for this test",
-)
-def test_renaming_top_level_directory_on_windows(
-    p: P,
-    event_queue: TestEventQueue,
-    start_watching: StartWatching,
-) -> None:
-    start_watching()
-
-    mkdir(p("a"))
-    event = event_queue.get(timeout=5)[0]
-    assert isinstance(event, DirCreatedEvent)
-    assert event.src_path == p("a")
-
-    mkdir(p("a", "b"))
-    event = event_queue.get(timeout=5)[0]
-    assert isinstance(event, DirCreatedEvent)
-    assert event.src_path == p("a", "b")
-
-    event = event_queue.get(timeout=5)[0]
-    assert isinstance(event, DirCreatedEvent)
-    assert event.src_path == p("a", "b")
-
-    event = event_queue.get(timeout=5)[0]
-    assert isinstance(event, DirModifiedEvent)
-    assert event.src_path == p("a")
-
-    mv(p("a"), p("a2"))
-    event = event_queue.get(timeout=5)[0]
-    assert isinstance(event, DirMovedEvent)
-    assert event.src_path == p("a", "b")
-
-    open(p("a2", "b", "c"), "a").close()
-
-    events = []
-    while True:
-        events.append(event_queue.get(timeout=5)[0])
-        if event_queue.empty():
-            break
-
-    assert all(
-        [
-            isinstance(
-                e, (FileCreatedEvent, FileMovedEvent, DirMovedEvent, DirModifiedEvent)
-            )
-            for e in events
-        ]
-    )
-
-    for event in events:
-        if isinstance(event, FileCreatedEvent):
-            assert event.src_path == p("a2", "b", "c")
-        elif isinstance(event, FileMovedEvent):
-            assert event.dest_path == p("a2", "b", "c")
-            assert event.src_path == p("a", "b", "c")
-        elif isinstance(event, DirMovedEvent):
-            assert event.dest_path == p("a2")
-            assert event.src_path == p("a")
-        elif isinstance(event, DirModifiedEvent):
-            assert event.src_path == p("a2", "b")
-
-
-@pytest.mark.skipif(
-    platform.is_windows(), reason="Windows create another set of events for this test"
-)
+@pytest.mark.skipif(platform.is_windows(), reason="Windows create another set of events for this test")
 def test_move_nested_subdirectories(
     p: P,
     event_queue: TestEventQueue,
@@ -610,8 +516,8 @@ def test_move_nested_subdirectories(
     expect_event(DirModifiedEvent(p("dir1")))
     expect_event(DirModifiedEvent(p()))
 
-    expect_event(DirMovedEvent(p("dir1", "dir2", "dir3"), p("dir2", "dir3")))
-    expect_event(FileMovedEvent(p("dir1", "dir2", "dir3", "a"), p("dir2", "dir3", "a")))
+    expect_event(DirMovedEvent(p("dir1", "dir2", "dir3"), p("dir2", "dir3"), is_synthetic=True))
+    expect_event(FileMovedEvent(p("dir1", "dir2", "dir3", "a"), p("dir2", "dir3", "a"), is_synthetic=True))
 
     if platform.is_bsd():
         event = event_queue.get(timeout=5)[0]
@@ -691,9 +597,7 @@ def test_move_nested_subdirectories_on_windows(
 
 
 @pytest.mark.flaky(max_runs=5, min_passes=1, rerun_filter=rerun_filter)
-@pytest.mark.skipif(
-    platform.is_bsd(), reason="BSD create another set of events for this test"
-)
+@pytest.mark.skipif(platform.is_bsd(), reason="BSD create another set of events for this test")
 def test_file_lifecyle(p: P, start_watching: StartWatching, expect_event: ExpectEvent) -> None:
     start_watching()